requirements.txt
numpy==1.21.5
django==4.1.4
pillow==9.0.1
joblib==1.2.0
progress==1.6
scikit-learn==1.1.3
